pyproject.toml
[project]
name = "aiagent-with-observability"
version = "0.1.0"
description = "AI Agent with comprehensive observability using LangGraph, CrewAI, and Langfuse"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai>=1.6.1",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.0",
    "langfuse>=3.10.3",
    "langgraph>=1.0.4",
    "litellm>=1.80.7",
    "openinference-instrumentation-crewai>=0.1.16",
    "python-dotenv>=1.2.1",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["aiagent*"]
